+++ src/configuration/__init__.py
@@ -4,6 +4,8 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
+    from zoneinfo import ZoneInfo
+
     from integrations.openwb.charging_station import ChargingStation
 
 
@@ -27,6 +29,7 @@ class Configuration:
         self.saic_tenant_id: str = "459771"
         self.saic_relogin_delay: int = 15 * 60  # in seconds
         self.saic_read_timeout: float = 10.0  # in seconds
+        self.saic_user_timezone: ZoneInfo | None = None
         self.battery_capacity_map: dict[str, float] = {}
         self.mqtt_host: str | None = None
         self.mqtt_port: int = 1883